Server now uses Asynchat, Working example.
So the Client now actually sends a request for the list of cities, it is
answered, client stores it, Client asks for work every 10 seconds when
there is no work to be done.
Server responds to requests with a fixed starting point for Greedy algo,
but the client proccesses it, and returns the whole route.  server
calculates length and displays it.
so functionality is working in the client, (though just for greedy with
start city) the server just needs logic added to rotate through cities,
and keep current best, etc.